+++ Engine/Render/_public.h
@@ -22,7 +22,7 @@ void MYX_UploadVisibleTilemap(const byte* tilemap, byte x, byte y, byte w);
 void MYX_SetSpritePalette(byte index, const void* data, byte count);
 void MYX_SetTilemapPalette(byte index, const void* data, byte count);
 
-void MYX_PutSprite(int x, byte y, MYXSprite sprite);
+void MYX_PutSprite(int x, int y, MYXSprite sprite);
 
 void MYX_SetTilemapOffset(byte x, byte y);
